Products

+++ composeApp/src/commonMain/kotlin/domain/Models/AvailableTime.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class AvailableTime(@SerialName("id") val id: Int? = null, @SerialName("time_id") val time: Int? = null,
+data class AvailableTime(@SerialName("id") val id: Long? = null, @SerialName("time_id") val time: Int? = null,
                          @SerialName("vendor_time") val vendorTime: VendorTime? = null, var isAvailable: Boolean = false,
                          val isSelected: Boolean = false): Parcelable
