+++ composeApp/src/commonMain/kotlin/domain/Models/ServiceTypeTherapists.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.Serializable
 
 
 @Serializable @Parcelize
-data class ServiceTypeTherapists(@SerialName("id") val id: Int? = null, @SerialName("therapist_id") val therapistId: Int? = null,
-                                 @SerialName("service_type_id") val serviceTypeId: Int? = null, @SerialName("therapist_info") val therapistInfo: TherapistInfo? = null,
+data class ServiceTypeTherapists(@SerialName("id") val id: Long? = null, @SerialName("therapist_id") val therapistId: Long? = null,
+                                 @SerialName("service_type_id") val serviceTypeId: Long? = null, @SerialName("therapist_info") val therapistInfo: TherapistInfo? = null,
                                  val isSelected: Boolean = false): Parcelable
 data class ServiceTypeTherapistUIModel(
     val selectedTherapist: ServiceTypeTherapists?,
